pyproject.toml
[project]
name = "kb"
dynamic = ["version"]
description = "LLM-powered personal knowledge base"
requires-python = ">=3.11"
dependencies = [
    "defusedxml>=0.7",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0",
]
web = [
    "requests>=2.31",
    "html2text>=2024.1",
]
pdf = [
    "pymupdf>=1.24",
]
ocr = [
    "pymupdf>=1.24",
    "pdf2image>=1.17",
    "pytesseract>=0.3",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.scripts]
kb = "kb.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kb"]

[tool.hatch.version]
path = "kb/__init__.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=kb --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["kb"]

[tool.coverage.report]
show_missing = true
skip_covered = false
